Move project component out of page
