+++ app/projects/page.tsx
@@ -1,79 +1,8 @@
-import {
-  IconName,
-  IconPrefix,
-  faArrowLeft,
-} from "@fortawesome/free-solid-svg-icons";
+import { faArrowLeft } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import projectsJSON from "./api/projects.json";
 import Project from "../types/project";
-import StyledMarkdown from "../shared-components/styled-markdown";
-
-interface ProjectComponentProps {
-  project: Project;
-}
-
-function ProjectComponent(props: ProjectComponentProps) {
-  const { title, longDescription, pointers, tags, image, links } =
-    props.project;
-
-  return (
-    <section className="flex flex-row p-8 mx-4">
-      <div className="flex flex-col min-w-[25%]">
-        <img src={image} className="object-cover max-w-[75%]" />
-      </div>
-      <div className="flex flex-col min-w-[75%] -mt-1">
-        <div className="flex flex-col space-y-6">
-          <div className="space-y-2">
-            <div className="flex flex-row space-x-3">
-              <h6 className="font-medium text-2xl">{title}</h6>
-              {links &&
-                links.map((link, index) => {
-                  return (
-                    <a
-                      key={index}
-                      href={link.url}
-                      target="_blank"
-                      className="my-auto hover:text-blue-400"
-                    >
-                      <FontAwesomeIcon
-                        icon={link.icon.split(" ") as [IconPrefix, IconName]}
-                        size="lg"
-                      />
-                    </a>
-                  );
-                })}
-            </div>
-            <StyledMarkdown>{longDescription}</StyledMarkdown>
-          </div>
-          <ul className="space-y-2 text-gray-300">
-            {pointers.map((pointer, index) => {
-              return (
-                <li key={index}>
-                  <div className="flex flex-row">
-                    <span className="ms-2 me-3">·</span>
-                    <StyledMarkdown>{pointer}</StyledMarkdown>
-                  </div>
-                </li>
-              );
-            })}
-          </ul>
-          <div className="flex flex-row space-x-4">
-            {tags.map((tag, index) => {
-              return (
-                <span
-                  key={index}
-                  className="bg-blue-600/70 py-1 px-2 text-sm text-blue-100"
-                >
-                  {tag}
-                </span>
-              );
-            })}
-          </div>
-        </div>
-      </div>
-    </section>
-  );
-}
+import ProjectComponent from "./_components/project";
 
 export default function Projects() {
   const projects = projectsJSON as Project[];
